// frontend/src/api/model/githubCom4H1RZooraInternalDomainAuditEntry.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Zoora API
 * REST API for the Zoora education platform.
 * OpenAPI spec version: 1.0
 */
import type { GithubCom4H1RZooraInternalDomainAuditAction } from './githubCom4H1RZooraInternalDomainAuditAction';
import type { GithubCom4H1RZooraInternalDomainAuditEntryMetadata } from './githubCom4H1RZooraInternalDomainAuditEntryMetadata';
import type { GithubCom4H1RZooraInternalDomainAuditOutcome } from './githubCom4H1RZooraInternalDomainAuditOutcome';
import type { GithubCom4H1RZooraInternalDomainAuditTargetType } from './githubCom4H1RZooraInternalDomainAuditTargetType';

export interface GithubCom4H1RZooraInternalDomainAuditEntry {
  action?: GithubCom4H1RZooraInternalDomainAuditAction;
  actor_id?: string;
  actor_name?: string;
  actor_username?: string;
  created_at?: string;
  id?: string;
  metadata?: GithubCom4H1RZooraInternalDomainAuditEntryMetadata;
  organization_id?: string;
  outcome?: GithubCom4H1RZooraInternalDomainAuditOutcome;
  target_id?: string;
  target_label?: string;
  target_type?: GithubCom4H1RZooraInternalDomainAuditTargetType;
}