// frontend/src/api/model/githubCom4H1RZooraInternalDomainAuditAction.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Zoora API
 * REST API for the Zoora education platform.
 * OpenAPI spec version: 1.0
 */

export type GithubCom4H1RZooraInternalDomainAuditAction = typeof GithubCom4H1RZooraInternalDomainAuditAction[keyof typeof GithubCom4H1RZooraInternalDomainAuditAction];


export const GithubCom4H1RZooraInternalDomainAuditAction = {
  AuditCreated: 'created',
  AuditUpdated: 'updated',
  AuditDeleted: 'deleted',
  AuditEnrolled: 'enrolled',
  AuditUnenrolled: 'unenrolled',
  AuditGraded: 'graded',
  AuditDisabled: 'disabled',
  AuditEnabled: 'enabled',
} as const;