// frontend/src/api/model/githubCom4H1RZooraInternalDomainAuditOutcome.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Zoora API
 * REST API for the Zoora education platform.
 * OpenAPI spec version: 1.0
 */

export type GithubCom4H1RZooraInternalDomainAuditOutcome = typeof GithubCom4H1RZooraInternalDomainAuditOutcome[keyof typeof GithubCom4H1RZooraInternalDomainAuditOutcome];


export const GithubCom4H1RZooraInternalDomainAuditOutcome = {
  AuditOutcomeSuccess: 'success',
  AuditOutcomeDenied: 'denied',
} as const;